+++ workers/unity/Assets/Playground/Editor/SnapshotGenerator/SnapshotGenerator.cs
@@ -98,7 +98,7 @@ namespace Playground.Editor.SnapshotGenerator
         {
             const string entityType = "Spinner";
 
-            var transform = TransformUtils.CreateTransformSnapshot(coords, Quaternion.identity);
+            var transform = TransformUtils.CreateTransformSnapshot(coords.ToUnityVector(), Quaternion.identity);
 
             var template = new EntityTemplate();
             template.AddComponent(new Position.Snapshot(coords), WorkerUtils.UnityGameLogic);